move retryConfiguration and ArrayDeserilization enum to client configuration & add remaining methods in CoreHelper from APIHelper.
